refactor(devtools): dedupe app smoke/test runners
